Wrap bootinfo, simplify header files dependencies
